Revert "feat(metric): adds accepted/closed connections counter metric for controller and worker clusters (#2593)"
This reverts commit bf7e236b8b824b94611e8874712d4fe8b4797de4.
