+++ internal/daemon/internal/metric/testing.go
