+++ internal/daemon/worker/internal/metric/cluster_client.go
@@ -2,9 +2,10 @@ package metric
 
 import (
 	"context"
+	"time"
 
 	"github.com/hashicorp/boundary/globals"
-	"github.com/hashicorp/boundary/internal/daemon/metric"
+	metric "github.com/hashicorp/boundary/internal/daemon/internal/metric"
 	"github.com/hashicorp/boundary/internal/gen/controller/servers/services"
 	"github.com/prometheus/client_golang/prometheus"
 	"google.golang.org/grpc"
@@ -28,6 +29,35 @@ var grpcRequestLatency prometheus.ObserverVec = prometheus.NewHistogramVec(
 	metric.ListGrpcLabels,
 )
 
+type requestRecorder struct {
+	reqLatency prometheus.ObserverVec
+	labels     prometheus.Labels
+
+	// measurements
+	start time.Time
+}
+
+// NewRequestRecorder creates a requestRecorder struct which is used to measure gRPC client request latencies.
+// For testing purposes, this method is exported.
+func newRequestRecorder(fullMethodName string, reqLatency prometheus.ObserverVec) requestRecorder {
+	service, method := metric.SplitMethodName(fullMethodName)
+	r := requestRecorder{
+		reqLatency: reqLatency,
+		labels: prometheus.Labels{
+			metric.LabelGrpcMethod:  method,
+			metric.LabelGrpcService: service,
+		},
+		start: time.Now(),
+	}
+
+	return r
+}
+
+func (r requestRecorder) Record(err error) {
+	r.labels[metric.LabelGrpcCode] = metric.StatusFromError(err).Code().String()
+	r.reqLatency.With(r.labels).Observe(time.Since(r.start).Seconds())
+}
+
 // The expected codes returned by the grpc client calls to cluster services.
 var expectedGrpcClientCodes = []codes.Code{
 	codes.OK, codes.Canceled, codes.Unknown, codes.InvalidArgument, codes.DeadlineExceeded, codes.NotFound,
@@ -41,7 +71,7 @@ var expectedGrpcClientCodes = []codes.Code{
 // between the cluster and its clients.
 func InstrumentClusterClient() grpc.UnaryClientInterceptor {
 	return func(ctx context.Context, method string, req, reply interface{}, cc *grpc.ClientConn, invoker grpc.UnaryInvoker, opts ...grpc.CallOption) error {
-		r := metric.NewGrpcRequestRecorder(method, grpcRequestLatency)
+		r := newRequestRecorder(method, grpcRequestLatency)
 		err := invoker(ctx, method, req, reply, cc, opts...)
 		r.Record(err)
 		return err
