+++ internal/daemon/worker/internal/metric/proxy_http.go
@@ -8,7 +8,7 @@ import (
 	"path"
 
 	"github.com/hashicorp/boundary/globals"
-	"github.com/hashicorp/boundary/internal/daemon/metric"
+	"github.com/hashicorp/boundary/internal/daemon/internal/metric"
 	"github.com/prometheus/client_golang/prometheus"
 	"github.com/prometheus/client_golang/prometheus/promhttp"
 )
